+++ CHANGELOG.md
@@ -5,7 +5,9 @@ Contributors: please follow the recommendations outlined at [keepachangelog.com]
 
 ## [Unreleased]
 *Please add entries here for your pull requests.*
-Add option to specify i18n_yml_dir in order to include only subset of locale files when generating translations.js & default.js for react-intl
+
+### Added
+- Add option to specify i18n_yml_dir in order to include only subset of locale files when generating translations.js & default.js for react-intl
 
 ## [6.8.2] - 2017-03-24
 ## Fixed