+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Contributors: please follow the recommendations outlined at [keepachangelog.com]
 ## Improved
 - Use <script type="application/json"> for props and store instead of hidden div [#775] (https://github.com/shakacode/react_on_rails/pull/775) by [cheremukhin23](https://github.com/cheremukhin23).
 
+### Added
+- Add option to specify i18n_yml_dir in order to include only subset of locale files when generating translations.js & default.js for react-intl
+[#777](https://github.com/shakacode/react_on_rails/pull/777) by [danijel](https://github.com/danijel).
+
 ## [6.8.2] - 2017-03-24
 ## Fixed
 - Change webpack output path to absolute and update webpack to version ^2.3.1. [#771](https://github.com/shakacode/react_on_rails/pull/771) by [cheremukhin23](https://github.com/cheremukhin23).