+++ CHANGELOG.md
@@ -5,10 +5,13 @@ Contributors: please follow the recommendations outlined at [keepachangelog.com]
 
 ## [Unreleased]
 
+## [6.9.2] - 2017-04-02
+
 ### Changed
 - Update version_checker.rb to `logger.error` rather than `logger.warn` for gem/npm version mismatch. [#788](https://github.com/shakacode/react_on_rails/issues/788) by [justin808](https://github.com/justin808).
 
 ### Fixed
+- Remove pretty formatting of JSON in development. [#789](https://github.com/shakacode/react_on_rails/pull/789) by [justin808](https://github.com/justin808)
 - Clear hydrated stores with each server rendered block. [#785](https://github.com/shakacode/react_on_rails/pull/785) by [udovenko](https://github.com/udovenko)
 
 ## [6.9.1] - 2017-03-30
@@ -515,7 +518,8 @@ Best done with Object destructing:
 ##### Fixed
 - Fix several generator related issues.
 
-[Unreleased]: https://github.com/shakacode/react_on_rails/compare/6.9.1...master
+[Unreleased]: https://github.com/shakacode/react_on_rails/compare/6.9.2...master
+[6.9.2]: https://github.com/shakacode/react_on_rails/compare/6.9.1...6.9.2
 [6.9.1]: https://github.com/shakacode/react_on_rails/compare/6.8.2...6.9.1
 [6.9.0]: https://github.com/shakacode/react_on_rails/compare/6.8.2...6.9.0
 [6.8.2]: https://github.com/shakacode/react_on_rails/compare/6.8.1...6.8.2